chore(eslint.yml): add ESLint workflow to run eslint scanning on push, pull request, and schedule
This commit adds a new file `.github/workflows/eslint.yml` which contains the configuration for an ESLint workflow. The workflow is triggered on push to the `main` branch, pull requests targeting the `main` branch, and on a schedule every Wednesday at 6:42 AM.

The workflow runs on the `ubuntu-latest` environment and has the necessary permissions to read repository contents and write security events. The steps in the workflow include checking out the code, installing ESLint version 8.10.0 and the `@microsoft/eslint-formatter-sarif` package, running ESLint with the specified configuration and output format, and uploading the analysis results to GitHub using the `github/codeql-action/upload-sarif` action.

The purpose of this workflow is to identify and report on patterns found in ECMAScript/JavaScript code using ESLint. The analysis results are uploaded to GitHub for further processing and review.