chore(eslint.yml): add ESLint workflow to run eslint scanning on push, pull request, and schedule
This commit adds a new file `.github/workflows/eslint.yml` which contains the configuration for an ESLint workflow. The workflow is triggered on push to the `main` branch, pull requests targeting the `main` branch, and on a schedule every Wednesday at 6:42 AM.

The workflow runs on the `ubuntu-latest` environment and has the necessary permissions to read the repository contents and write security events. It consists of the following steps:

1. Checkout code: This step uses the `actions/checkout` action to fetch the repository code.
2. Install ESLint: This step installs ESLint version 8.10.0 and the `@microsoft/eslint-formatter-sarif` package.
3. Run ESLint: This step runs ESLint using the specified configuration file `.eslintrc.js` and formats the output using the `@microsoft/eslint-formatter-sarif`. The results are saved to the `eslint-results.sarif` file. The `continue-on-error` flag is set to true to allow the workflow to continue even if there are ESLint errors.
4. Upload analysis results to GitHub: This step uses the `github/codeql-action/upload-sarif` action to upload the generated SARIF file to GitHub for further analysis. The `wait-for-processing` flag is set to true to wait for the analysis results to be processed.

This workflow helps identify and report patterns found in ECMAScript/JavaScript code using ESLint.